walker: replace filepath.Walk with fs.WalkDir
As per Go 1.16 release notes https://go.dev/doc/go1.16#path/filepath

> The new function `WalkDir` is similar to `Walk`, but is typically more efficient.
> The function passed to `WalkDir` receives a `fs.DirEntry` instead of a `fs.FileInfo`.

This makes testing of the walker easier with testing/fstest,
or any other fs.FS compatible libraries.

Another side-effect is that walker is now aligned with the FS interpretation
of the rest of the codebase in that open documents are served from in-memory cache, not from disk.
This may have another positive impact on performance.